move subgraph building tools from mock-block-dock to @blockprotocol/graph
