+++ packages/@blockprotocol/graph/src/internal.ts
@@ -1,0 +1,4 @@
+export {
+  addEntitiesToSubgraphByMutation,
+  addKnowledgeGraphEdgeToSubgraphByMutation,
+} from "./internal/mutate-subgraph";
