+++ packages/@blockprotocol/graph/src/internal/mutate-subgraph.ts
@@ -7,12 +7,19 @@ import {
   KnowledgeGraphVertices,
   Subgraph,
   Timestamp,
-} from "@blockprotocol/graph";
+} from "../types.js";
+import { isEqual } from "./mutate-subgraph/is-equal.js";
 
-import { isEqual } from "../util";
-
-/** @todo - clean up the assertions here */
-export const addKnowledgeGraphEdge = (
+/**
+ * Looking to build a subgraph? You probably want `import { buildSubgraph } from @blockprotocol/graph/stdlib`
+ *
+ * This MUTATES the given `Subgraph` by adding the given outwardEdge to the entity at the specified atTime.
+ * @param subgraph – the subgraph to mutate by adding the outward edge
+ * @param sourceEntityId – the id of the entity the edge is coming from
+ * @param atTime – the time at which the edge should be recorded as being added at
+ * @param outwardEdge – the edge itself
+ */
+export const addKnowledgeGraphEdgeToSubgraphByMutation = (
   subgraph: Subgraph,
   sourceEntityId: EntityId,
   atTime: Timestamp,
@@ -28,10 +35,6 @@ export const addKnowledgeGraphEdge = (
     subgraph.edges[sourceEntityId]![atTime] = [outwardEdge];
   } else {
     const outwardEdgesAtTime = subgraph.edges[sourceEntityId]![atTime]!;
-    /**
-     * @todo - Q for PR review: Added a lodash dependency for this, equality of the outward edge is actually complicated
-     *    fine to keep?
-     */
     if (
       !outwardEdgesAtTime.find((otherOutwardEdge) =>
         isEqual(otherOutwardEdge, outwardEdge),
@@ -45,15 +48,17 @@ export const addKnowledgeGraphEdge = (
 };
 
 /**
- * Mutates the given `Subgraph` by adding the given entities into the `Vertices` object, creating edges as necessary.
+ * Looking to build a subgraph? You probably want `import { buildSubgraph } from @blockprotocol/graph/stdlib`
+ *
+ * This MUTATES the given `Subgraph` by adding the given entities into the `Vertices` object, creating edges as necessary.
  *
  * *Note*: This only adds edges as implied by the given entities, if the `Subgraph` is invalid at the time of method
  * call (e.g. by missing link endpoints), this will not loop through the vertex set to finish incomplete edges.
  *
- * @param subgraph
- * @param entities
+ * @param subgraph – the subgraph to mutate by adding the provided entities
+ * @param entities – the entity to add to the provided subgraph
  */
-export const addEntitiesToSubgraph = (
+export const addEntitiesToSubgraphByMutation = (
   subgraph: Subgraph,
   entities: Entity[],
 ) => {
@@ -99,26 +104,46 @@ export const addEntitiesToSubgraph = (
     linkEntityId,
     { leftEntityId, rightEntityId, earliestTime },
   ] of Object.entries(linkMap)) {
-    addKnowledgeGraphEdge(subgraph, linkEntityId, earliestTime, {
-      kind: "HAS_LEFT_ENTITY",
-      reversed: false,
-      rightEndpoint: { baseId: leftEntityId, timestamp: earliestTime },
-    });
-    addKnowledgeGraphEdge(subgraph, leftEntityId, earliestTime, {
-      kind: "HAS_LEFT_ENTITY",
-      reversed: true,
-      rightEndpoint: { baseId: linkEntityId, timestamp: earliestTime },
-    });
-    addKnowledgeGraphEdge(subgraph, linkEntityId, earliestTime, {
-      kind: "HAS_RIGHT_ENTITY",
-      reversed: false,
-      rightEndpoint: { baseId: rightEntityId, timestamp: earliestTime },
-    });
-    addKnowledgeGraphEdge(subgraph, rightEntityId, earliestTime, {
-      kind: "HAS_RIGHT_ENTITY",
-      reversed: true,
-      rightEndpoint: { baseId: linkEntityId, timestamp: earliestTime },
-    });
+    addKnowledgeGraphEdgeToSubgraphByMutation(
+      subgraph,
+      linkEntityId,
+      earliestTime,
+      {
+        kind: "HAS_LEFT_ENTITY",
+        reversed: false,
+        rightEndpoint: { baseId: leftEntityId, timestamp: earliestTime },
+      },
+    );
+    addKnowledgeGraphEdgeToSubgraphByMutation(
+      subgraph,
+      leftEntityId,
+      earliestTime,
+      {
+        kind: "HAS_LEFT_ENTITY",
+        reversed: true,
+        rightEndpoint: { baseId: linkEntityId, timestamp: earliestTime },
+      },
+    );
+    addKnowledgeGraphEdgeToSubgraphByMutation(
+      subgraph,
+      linkEntityId,
+      earliestTime,
+      {
+        kind: "HAS_RIGHT_ENTITY",
+        reversed: false,
+        rightEndpoint: { baseId: rightEntityId, timestamp: earliestTime },
+      },
+    );
+    addKnowledgeGraphEdgeToSubgraphByMutation(
+      subgraph,
+      rightEntityId,
+      earliestTime,
+      {
+        kind: "HAS_RIGHT_ENTITY",
+        reversed: true,
+        rightEndpoint: { baseId: linkEntityId, timestamp: earliestTime },
+      },
+    );
   }
   /* eslint-enable no-param-reassign */
 };
